Bias fix and code cleanup
-Code cleanup
-Bias fix:

Now if you either specity a champion or a spatula, the program will score comps that are closer to the one you specified, meaning it will no longer only give you champions that are not related to the one you are interested in. It's an increase in the quality of the result!